refactor code, + add storage request (syncronous) calls. these are used to interaaaact direct with storage engine.
